% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunked.R
\name{assert_integer_ge_1}
\alias{assert_integer_ge_1}
\title{Assert number is an integer that is greater than or equal to 1}
\usage{
assert_integer_ge_1(x, arg_name)
}
\arguments{
\item{x}{An integer >= 1. Raises an error if this condition is not met}

\item{arg_name}{character. The argument name for x. This is used to generate
an informative error message.}
}
\description{
Helper function for \code{\link{fread_chunked}} and
\code{\link{process_df_chunked}}.
}
\seealso{
\code{\link{fread_chunked}}, \code{\link{process_df_chunked}}
}
