% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_cols_exist}
\alias{check_required_cols_exist}
\title{Check required columns are present}
\usage{
check_required_cols_exist(df, ...)
}
\arguments{
\item{df}{Dataframe}

\item{...}{Character vector(s)}
}
\description{
Combines supplied character vectors into a single vector, then checks whether
these are all present in `names(df)`. Raises an error if not.
}
