% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenodataset.R
\name{colname_to_field_inst_array_df}
\alias{colname_to_field_inst_array_df}
\title{Convert descriptive colnames to dataframe of FieldID/instance/array
combinations}
\usage{
colname_to_field_inst_array_df(x)
}
\arguments{
\item{x}{a character vector of descriptive colnames generated by
\code{\link{pheno_data_dict}}}
}
\value{
a dataframe of extracted FieldID/instance/array values and all their
  possible combinations
}
\description{
Convert descriptive colnames to dataframe of FieldID/instance/array
combinations
}
