% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_list_tables}
\alias{db_list_tables}
\title{Make a named list of \code{tbl_dbi} objects from a \code{DBIConnection}
object}
\usage{
db_list_tables(conn)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by
\code{\link[DBI]{dbConnect}}.}
}
\value{
A named list of \code{tbl_dbi} objects
}
\description{
A convenience function that returns a list of \code{tbl_dbi}
objects from a \code{DBIConnection} object, named as per the database table
names.
}
\details{
See the \href{https://dbplyr.tidyverse.org/articles/dbplyr.html}{introduction
to dbplyr} vignette for getting started with databases and
\code{\link[dplyr]{dplyr}}.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")
dplyr::copy_to(con, head(iris), "iris_head")
dplyr::copy_to(con, head(mtcars), "mtcars_head")
db_tables <- make_list_of_tbls_from_db(con)
db_tables$iris_head
db_table$mtcars_head
}
