% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{db_tables_to_list}
\alias{db_tables_to_list}
\title{Make a named list of \code{tbl_dbi} objects from a \code{DBIConnection}
object}
\usage{
db_tables_to_list(conn)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by
\code{\link[DBI]{dbConnect}}.}
}
\value{
A named list of \code{tbl_dbi} objects
}
\description{
A convenience function that returns a list of \code{tbl_dbi}
objects from a \code{DBIConnection} object, named as per the database table
names.
}
\details{
See the \href{https://dbplyr.tidyverse.org/articles/dbplyr.html}{introduction
to dbplyr} vignette for getting started with databases and
\code{\link[dplyr]{dplyr}}.
}
\examples{
# create an SQLite database in memory
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")

# populate with tables
dplyr::copy_to(con, head(iris), "iris_head")
dplyr::copy_to(con, head(mtcars), "mtcars_head")

# create named list of tbl_dbi objects
db_tables <- db_tables_to_list(con)

# view tables
db_tables$iris_head
db_tables$mtcars_head

# to import to R, use dplyr::collect()
dplyr::collect(db_tables$iris_head)
}
