% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{derive_dob}
\alias{derive_dob}
\title{Derive an estimated participant date of birth}
\usage{
derive_dob(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  .drop = FALSE,
  .details_only = FALSE
)
}
\arguments{
\item{ukb_main}{A UK Biobank main dataset.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
 showcase}. This should be a data frame where all columns are of type
\code{character}. By default, this is downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{.drop}{If \code{TRUE}, remove the required input columns from the
result}

\item{.details_only}{If \code{TRUE}, return a list containing details of
required input variables (Field IDs) and derived variables (new column
name, label and values/value labels).}
}
\value{
A data frame with a column called \code{dob} (type
  character).
}
\description{
Estimates dates of birth from year and month of birth (Field IDs
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=34}{34} and
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=52}{52} respectively) as
the first date of the month.
}
\examples{
ukb_main <- data.frame(eid = 1, f34_0_0 = 1990, f52_0_0 = 1)
 # keep input year/month of birth columns
 derive_dob(ukb_main)

 # remove input year/month of birth columns
 derive_dob(ukb_main, .drop = TRUE)
}
