% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_algorithms.R
\name{diabetes_diagnosis_date_first_occurrence}
\alias{diabetes_diagnosis_date_first_occurrence}
\title{Get date of diabetes diagnosis based on first occurrence fields}
\usage{
diabetes_diagnosis_date_first_occurrence(ukb_pheno, data_dict)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}
}
\value{
The input UK Biobank dataset with an additional column for date of
  diabetes diagnosis based on first occurrence fields
}
\description{
Wrapper around \code{\link{rowise_min_max_date}}. Extracts the earliest date
across FieldID's 130706, 130708, 130712, 130710, 130714, 132202. See
documentation for \code{\link{diabetes_type_first_occurrence}} also.
}
