% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_dummy_ukb_data_to_tempdir}
\alias{download_dummy_ukb_data_to_tempdir}
\title{Download dummy UK Biobank data}
\usage{
download_dummy_ukb_data_to_tempdir()
}
\value{
The path to the downloaded file.
}
\description{
A dummy UKB data file is downloaded to \code{\link[base]{tempdir}}. This will
be removed at the end of the current R session. The path to this file is
returned as a string.
}
\examples{

\dontrun{
# download dummy data to tempdir() and get filepath
# dummy_ukb_data_path <- download_dummy_ukb_data_to_tempdir()

# make data dictionary make_data_dict(dummy_ukb_data_path, delim = ",",
# ukb_data_dict = get_ukb_data_dict())
}
}
