% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_pheno_summaries.R
\name{dt_rowwise_fn}
\alias{dt_rowwise_fn}
\title{Helper function for summarise_rowise()}
\usage{
dt_rowwise_fn(fn, cols, ...)
}
\arguments{
\item{fn}{a character vector of functions}

\item{cols}{a character vector of column names to summarise with functions in
fn}

\item{...}{allows additional arguments such as na.rm to passed on to
functions listed in fn}
}
\description{
Helper function for summarise_rowise()
}
