% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_main_dataset_clinical_events}
\alias{dummy_main_dataset_clinical_events}
\title{Example clinical events data from a UK Biobank main dataset}
\usage{
dummy_main_dataset_clinical_events()
}
\value{
A data frame of dummy data
}
\description{
Example dummy clinical events data that can be tidied with \code{\link{tidy_clinical_events}}
}
\examples{
dummy_main_dataset_clinical_events()
}
\seealso{
\code{\link{tidy_clinical_events}}
}
