% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_events.R
\name{example_clinical_codes}
\alias{example_clinical_codes}
\title{Generate an example data frame of clinical codes for diabetes}
\usage{
example_clinical_codes()
}
\value{
A data frame of selected clinical codes for diabetes.
}
\description{
Data frames of this format may be used with \code{\link{extract_phenotypes}}.
Use \code{\link{validate_clinical_codes}} to check whether a clinical codes
data frame meets all requirements.
}
\details{
Note: this example is not an exhaustive list of codes for diabetes.
}
\examples{
example_clinical_codes()
}
\seealso{
\code{\link{validate_clinical_codes}}
}
