% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_codings_for_fieldid}
\alias{extract_codings_for_fieldid}
\title{Get UKB codings for a FieldID}
\usage{
extract_codings_for_fieldid(field_id, ukb_data_dict, ukb_codings)
}
\arguments{
\item{field_id}{Character - UKB Field ID}

\item{ukb_data_dict}{UKB data dictionary}

\item{ukb_codings}{UKB Codings dictionary}
}
\description{
General helper function
}
