% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_codings_for_fieldids}
\alias{extract_codings_for_fieldids}
\title{Get UKB codings for one or more FieldIDs}
\usage{
extract_codings_for_fieldids(field_ids, ukb_data_dict, ukb_codings)
}
\arguments{
\item{field_ids}{Character. One or more UKB FieldIDs}

\item{ukb_data_dict}{UKB data dictionary}

\item{ukb_codings}{UKB Codings dictionary}
}
\value{
Dataframe.
}
\description{
General helper function - filters
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{`ubk_codings`}
for the codings associated with one or more FieldIDs.
}
