% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{extract_first_or_last_clinical_event_multi}
\alias{extract_first_or_last_clinical_event_multi}
\title{Get the earliest record date for multiple phenotype categories}
\usage{
extract_first_or_last_clinical_event_multi(
  df,
  clinical_codes_df,
  min_max = "min",
  prefix = NULL,
  workers = NULL,
  df_pre_filtering_fn = NULL
)
}
\arguments{
\item{df}{Either a dataframe or a \code{\link[dbplyr]{tbl_dbi}} object generated
by one of the \code{get_XXX_diagnoses} functions}

\item{clinical_codes_df}{data frame. Must match the format as per
\code{\link{generate_self_reported_diabetes_codes_df}}.}

\item{min_max}{can be either "min" or "max". Extracts either the earliest or
latest date that on of the specified \code{codes} appears.}

\item{prefix}{character. Optionally add a prefix to column names.}

\item{workers}{integer. The number of processes to run in parallel when
\code{clinical_codes_df} includes multiple diseases. This is used to set
the number of workers in \code{\link[future]{plan}} with \code{strategy =
\link[future]{multisession}}, which is passed on to
\code{\link[furrr]{future_map}}. Default value is \code{NULL}.}

\item{df_pre_filtering_fn}{function. Optionally supply a function to
pre-filter \code{df}.For example, filter for only self-reported clinical
events with 'function(df) df %>% dplyr::filter(source == "f20002")'.
The function should only have one argument e.g. 'df'. This is designed to
enable parallel processing when \code{df} is a
\code{\link[dbplyr]{tbl_dbi}} object. Default value is \code{NULL}.}
}
\value{
A named list of data frames, one for each disease. Each data frame
  has an "eid" column, and "event_min/max_indicator" and "event_min/max_date"
  columns for each phenotype in the 'category' column of
  \code{clinical_codes_df} for that disease.
}
\description{
Builds on \code{\link{extract_first_or_last_clinical_event}}, extracting
either the first or last recorded clinical event for multiple phenotype
categories.
}
\seealso{
Other extract specific diagnostic codes functions: 
\code{\link{extract_first_or_last_clinical_event}()},
\code{\link{extract_single_diagnostic_code_record_basis}()}
}
\concept{extract specific diagnostic codes functions}
