% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{field_id_pivot_longer}
\alias{field_id_pivot_longer}
\title{Convert column for a FieldID to long format}
\usage{
field_id_pivot_longer(ukb_pheno, field_id, data_dict, ukb_codings)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{field_id}{Character. A UK Biobank Field ID}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}
}
\value{
A dataframe with 3 columns: \itemize{ \item eid \item column names
  (labelled as the specified `field_id`, prefixed by 'f') \item column values
  (labelled as he specified `field_id`, prefixed by 'f' and suffixed by
  '_value') }
}
\description{
Selects the columns relating to a specified FieldID (plus the 'eid' column)
and applies \code{\link[dplyr]{pivot_longer}}.
}
