% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{field_id_pivot_longer_multi}
\alias{field_id_pivot_longer_multi}
\title{Apply pivot_longer across multiple sets of columns}
\usage{
field_id_pivot_longer_multi(field_ids, ukb_pheno, data_dict, ukb_codings)
}
\arguments{
\item{field_ids}{Character vector of FieldIDs}

\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}
}
\value{
Dataframe.
}
\description{
Helper function.
}
\details{
This should only be applied to related FieldIDs which are each associated
with the same number of columns. For example, FieldIDs 20002, 20008, and
20009 all pertain to self-reported non-cancer illnesses.

Raises and error if the supplied FieldIDs are associated with differing
numbers of columns, or if the result has more rows than expected.
}
\section{Under the hood}{


  Maps \code{\link{field_id_pivot_longer}} to a set of FieldIDs, then
  recombines the results into a single dataframe using
  \code{\link[dplyr]{inner_join}}.
}

\seealso{
Other extract disease outcomes helpers: 
\code{\link{field_id_pivot_longer}()}
}
\concept{extract disease outcomes helpers}
