% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_data_dict}
\alias{filter_data_dict}
\title{Generic helper function values from a data dictionary}
\usage{
filter_data_dict(data_dict, filter_col, filter_value, return_col)
}
\arguments{
\item{data_dict}{a data dictionary generated by \code{\link{make_data_dict}}}

\item{filter_col}{Character. The column in `data_dict` to filter on.}

\item{filter_value}{Character (single or multiple). Filter for values in
`filter_col` that match one of these.}

\item{return_col}{Character. The column in `data_dict` to return after
filtering.}
}
\value{
A vector.
}
\description{
Filters a UK Biobank data dictionary on one column and returns the values
from another column (e.g. filter data dictionary for a FieldID, and return
the corresponding descriptive column name(s))
}
\details{
General helper function. Raises an error if produces and empty vector
}
\seealso{
Other data dictionary helper functions: 
\code{\link{get_colnames_for_fieldids}()}
}
\concept{data dictionary helper functions}
