% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenodataset.R
\name{format_ukb_df_header}
\alias{format_ukb_df_header}
\title{Processes a ukb pheno file header to match the form 'f.5912.0.0'}
\usage{
format_ukb_df_header(colheaders)
}
\arguments{
\item{colheaders}{A character vector.}
}
\value{
A ukb pheno file header (character vector) of the form 'f.5912.0.0'.
  Returns the header unaltered if already in this form

Helper function for \code{pheno_data_dict()}
}
\description{
Processes a ukb pheno file header to match the form 'f.5912.0.0'
}
