% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fread_tsv_as_character}
\alias{fread_tsv_as_character}
\title{Read a tsv file with all columns as type character}
\usage{
fread_tsv_as_character(...)
}
\arguments{
\item{...}{additional arguments are passed on to \code{\link[data.table]{fread}}}
}
\description{
Wrapper around \code{\link[data.table]{fread}}
}
