% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{generate_self_reported_diabetes_codes_df}
\alias{generate_self_reported_diabetes_codes_df}
\title{Generate a data frame of clinical codes for self-reported diabetes in the UK
Biobank}
\usage{
generate_self_reported_diabetes_codes_df()
}
\value{
Data frame
}
\description{
Data frames of this format may be used with
\code{\link{extract_first_or_last_clinical_event_multi}} to facilitate
identification of clinical events (e.g. date of diabetes diagnosis).
}
\examples{
generate_self_reported_diabetes_codes_df()
}
