% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_cancer_register_icd10_diagnoses}
\alias{get_cancer_register_icd10_diagnoses}
\title{Get cancer register ICD10 diagnoses}
\usage{
get_cancer_register_icd10_diagnoses(ukb_pheno, data_dict, ukb_codings)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}
}
\value{
Dataframe
}
\description{
Returns a long format dataframe with all ICD10 cancer register diagnoses
with associated dates for each UK Biobank participant.
}
\details{
Reformats the data for FieldIDs 40006 and 40005 (cancer register diagnoses and dates, see
\href{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=100092}{category 100092}).
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()}
}
\concept{get all diagnostic codes}
