% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colname}
\alias{get_colname}
\title{Get associated single decriptive colname for a FieldID}
\usage{
get_colname(field_id, ukb_mapping_df)
}
\arguments{
\item{field_id}{A UK Biobank Field ID as a string}

\item{ukb_mapping_df}{a data dictionary generated by
\code{\link{pheno_data_dict}}}
}
\value{
A single descriptive colname as a string
}
\description{
Use this in functions which manipulate a UKB phenotype dataset processed by
\code{\link{read_pheno}}. Should the convention for descriptive column names
change then so will these functions, however changes would only need to be
updated at the start of each function.
}
