% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colnames_for_fieldid}
\alias{get_colnames_for_fieldid}
\title{Get associated single decriptive colname for a FieldID}
\usage{
get_colnames_for_fieldid(field_id, data_dict)
}
\arguments{
\item{field_id}{Character. A UK Biobank Field ID}

\item{data_dict}{a data dictionary generated by
\code{\link{pheno_data_dict}}}
}
\value{
A character vector of descriptive colnames
}
\description{
Use this in functions which manipulate a UKB phenotype dataset processed by
\code{\link{read_pheno}}. Should the convention for descriptive column names
change then so will these functions, however changes would only need to be
updated at the start of each function.
}
\seealso{
Other Data dictionary functions: 
\code{\link{filter_data_dict}()}
}
\concept{Data dictionary functions}
