% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colnames_for_fieldids}
\alias{get_colnames_for_fieldids}
\title{Get descriptive colnames associated with one or more FieldIDs}
\usage{
get_colnames_for_fieldids(
  field_ids,
  data_dict,
  scalar_output = FALSE,
  error_if_missing = TRUE
)
}
\arguments{
\item{field_ids}{character vector of UK Biobank Field IDs.}

\item{data_dict}{a data dictionary generated by \code{\link{make_data_dict}}}

\item{scalar_output}{bool. If \code{TRUE}, error raised if more than one
colname is returned. Default value is \code{FALSE}.}

\item{error_if_missing}{bool. Raise error if any values in
\code{filter_value} are missing. Default is \code{TRUE}.}
}
\value{
A character vector of descriptive colnames
}
\description{
Returns all descriptive column names matching one or more FieldIDs.
}
\details{
Use this in functions which manipulate a UKB phenotype dataset processed by
\code{\link{read_ukb}}. Should the convention for descriptive column names
change then so will these functions, however changes would only need to be
updated at the start of each function.
}
\concept{data dictionary helper functions}
