% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_diagnoses_basis_fn}
\alias{get_diagnoses_basis_fn}
\title{Melt diagnoses code and date columns in a main UKB dataset}
\usage{
get_diagnoses_basis_fn(
  ukb_pheno,
  data_dict,
  ukb_codings,
  code_col_field_id,
  date_col_field_id
)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset. Must be a data table}

\item{data_dict}{A data dictionary specific to the UKB phenotype file,
generated by \code{\link{make_data_dict}}.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.}

\item{code_col_field_id}{character. The Field ID representing 'code'
variables}

\item{date_col_field_id}{character. The Field ID representing 'date'
variables that correspond to \code{code_col_field_id}.}
}
\value{
A data table with 'eid', 'source' (Field ID for the code
column), 'code' and 'date' columns.
}
\description{
For most clinical events variables in the main UKB dataset there is a
diagnostic code variable and an associated date of event variable. There are
spread over several instances/arrays. This function melts these to long
format.
}
