% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_diagnoses_set_index_code_date_cols}
\alias{get_diagnoses_set_index_code_date_cols}
\title{Standardise a long format diagnoses dataframe}
\usage{
get_diagnoses_set_index_code_date_cols(
  get_clinical_events_df,
  index_col,
  code_col,
  date_col,
  remove_special_dates = FALSE
)
}
\arguments{
\item{get_clinical_events_df}{a dataframe created by a 'get_X_diagnoses' function}

\item{index_col}{Column indicating diagnosis source (e.g. HES, self-report)}

\item{code_col}{Column of codes}

\item{date_col}{Column of dates when code was recorded}

\item{remove_special_dates}{Logical. Remove 'special' date values if
requested. Default is \code{FALSE}, as only the self-report fields contain
such values.}
}
\description{
Helper function
}
\details{
Selects essential columns: eid, field_id, code and date recorded.
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_all_diagnostic_codes_multi}()},
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()},
\code{\link{main_dataset_diagnoses_to_sqlite_db}()}
}
\concept{get all diagnostic codes}
