% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_first_diagnostic_code_record}
\alias{get_first_diagnostic_code_record}
\title{Get the earliest record date for a set of diagnostic codes}
\usage{
get_first_diagnostic_code_record(df, codes)
}
\arguments{
\item{df}{A dataframe generated by one of the 'get all diagnostic codes'
functions}

\item{codes}{Character vector of diagnostic codes}
}
\value{
Dataframe
}
\description{
For a set of diagnostic codes, extract individuals with at least one of these
and retrieve the earliest recorded date (if available).
}
\section{Under the hood}{


  Filters a long format dataframe of diagnostic codes generated by one of the
  'get all diagnostic codes' functions (e.g.
  \code{\link{get_hes_icd10_diagnoses}}), filters this for the specified
  `codes` then groups by individual (eid). For each individual, the earliest
  date is then retrieved. If no dates are available, the first diagnostic
  code as it appears in the data is retrieved instead.
}

\seealso{
Other get specific diagnostic codes, first recorded: 
\code{\link{get_first_diagnostic_code_record_basis}()}
}
\concept{get specific diagnostic codes, first recorded}
