% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_first_diagnostic_code_record_basis}
\alias{get_first_diagnostic_code_record_basis}
\title{Get first record for a set of diagnostic codes (generic)}
\usage{
get_first_diagnostic_code_record_basis(df, codes)
}
\arguments{
\item{df}{A dataframe generated by one of the 'get all diagnostic codes'
functions}

\item{codes}{Character vector of diagnostic codes}
}
\value{
A dataframe nested by eid, with columns for the earliest date one of
  the specified codes was recorded
}
\description{
Generic helper function. For a set of diagnostic codes, extract eids with at
least one of these and the associated earliest recorded date.
}
\seealso{
Other get specific diagnostic codes, first recorded: 
\code{\link{get_first_diagnostic_code_record}()}
}
\concept{get specific diagnostic codes, first recorded}
