% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_hes_icd9_diagnoses}
\alias{get_hes_icd9_diagnoses}
\title{Get hospital inpatient ICD9 diagnoses}
\usage{
get_hes_icd9_diagnoses(ukb_pheno, data_dict, ukb_codings)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{remove_special_dates}{Set special date values to \code{NA}. Default it
\code{TRUE}.}
}
\value{
Dataframe
}
\description{
Returns a long format dataframe with all hospital inpatient ICD9 codes and
associated dates for each UK Biobank participant.
}
\details{
Reformats the data for FieldIDs 41271 and 41281 (hospital inpatient diagnoses
and dates, see
\href{https://biobank.ndph.ox.ac.uk/ukb/label.cgi?id=2002}{category 2002}).

Also removes 'special' date values by default (see \href{}{coding })
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_all_diagnostic_codes_multi}()},
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()},
\code{\link{main_dataset_diagnoses_to_sqlite_db}()}
}
\concept{get all diagnostic codes}
