% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_hes_opcs3_operations}
\alias{get_hes_opcs3_operations}
\title{Get hospital inpatient OPCS3 operative procedures}
\usage{
get_hes_opcs3_operations(ukb_pheno, data_dict, ukb_codings)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset. Must be a data table}

\item{data_dict}{a data dictionary for \code{ukb_pheno}. Can be a "filtered"
version including only the columns to be summarised.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.}
}
\value{
Data frame
}
\description{
Returns a long format dataframe with all summary hospital inpatient OPCS3
codes and associated dates for each UK Biobank participant.
}
\details{
Reformats the data for FieldIDs 41273 and 41283 (hospital inpatient OPCS3
operative procedures and dates, see
\href{https://biobank.ndph.ox.ac.uk/ukb/label.cgi?id=2005}{category 2005}).
}
\seealso{
Other get all diagnostic codes: 
\code{\link{OLD_get_cancer_register_icd10_diagnoses}()},
\code{\link{OLD_get_cancer_register_icd9_diagnoses}()},
\code{\link{OLD_get_death_data_icd10_diagnoses}()},
\code{\link{OLD_get_diagnoses_set_index_code_date_cols}()},
\code{\link{OLD_get_hes_icd10_diagnoses}()},
\code{\link{OLD_get_hes_icd9_diagnoses}()},
\code{\link{OLD_get_self_report_cancer_diagnoses}()},
\code{\link{OLD_get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{OLD_get_self_report_non_cancer_diagnoses}()},
\code{\link{get_all_diagnostic_codes_multi}()},
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_hes_opcs4_operations}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()},
\code{\link{main_dataset_diagnoses_to_sqlite_db}()}
}
\concept{get all diagnostic codes}
