% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_self_report_non_cancer_diagnoses}
\alias{get_self_report_non_cancer_diagnoses}
\title{Get self-reported non-cancer diagnoses}
\usage{
get_self_report_non_cancer_diagnoses(
  ukb_pheno,
  data_dict,
  ukb_codings,
  remove_special_dates = TRUE
)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{remove_special_dates}{Logical. Remove 'special' date values if
requested. Default is \code{TRUE}. See
\href{https://biobank.ndph.ox.ac.uk/ukb/coding.cgi?id=13}{data coding 13}.}
}
\value{
Dataframe
}
\description{
Returns a long format dataframe with all self-reported non-cancer diagnoses
with associated dates for each UK Biobank participant.
}
\details{
Reformats the data for FieldIDs 20002 and 20008 (self-reported non-cancer
diagnoses and dates, see
\href{https://biobank.ndph.ox.ac.uk/ukb/label.cgi?id=100074}{category 100074}). Coded using
\href{https://biobank.ndph.ox.ac.uk/showcase/coding.cgi?id=6}{data coding 6}.
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_all_diagnostic_codes_multi}()},
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{main_dataset_diagnoses_to_sqlite_db}()}
}
\concept{get all diagnostic codes}
