% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_code_mappings}
\alias{get_ukb_code_mappings}
\title{Get UK Biobank clinical code mappings file}
\usage{
get_ukb_code_mappings()
}
\value{
A named list.
}
\description{
Downloads the UK Biobank code mappings file
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{resource 592})
from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{ukbwranglr_resources}
github repo. The raw file is a large excel spreadsheet. This has been saved
in \code{.Rdata} format in
\href{https://github.com/rmgpanw/ukbwranglr_resources}{ukbwranglr_resources}
as a named list of data frames, one for each sheet in the original file.
}
\details{
This function downloads the \code{ukb_code_mappings.Rdata} file from
\href{https://github.com/rmgpanw/ukbwranglr_resources}{ukbwranglr_resources}
to a temporary directory before loading and returning the result.

\strong{Note:} This is a large object (>450 MB)
}
