% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_code_mappings_direct}
\alias{get_ukb_code_mappings_direct}
\title{Get UK Biobank clinical code mappings file directly from UKB website}
\usage{
get_ukb_code_mappings_direct()
}
\value{
A named list.
}
\description{
Downloads the UK Biobank code mappings file (\code{all_lkps_maps_v2.xlsx},
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{resource 592})
directly from the UKB website to a temporary directory at
\code{\link[base]{tempdir}}. This is then read into R as a named list of data
frames, one for each sheet in the original file.
}
\details{
\strong{Note:} This is a large object (>450 MB)
}
\seealso{
\code{\link{get_ukb_code_mappings}}
}
