% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_codings}
\alias{get_ukb_codings}
\title{Download UKB codings file directly from UKB website}
\usage{
get_ukb_codings(path = file.path(tempdir(), "Codings.tsv"))
}
\arguments{
\item{path}{Path where file will be downloaded to. If a file already exists
at this path, then this will be read into R without re-downloading.}
}
\value{
A data frame.
}
\description{
Downloads the UK Biobank codings list from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UK
Biobank website} and reads into R with all columns as character type.
}
\examples{
\dontrun{ get_ukb_codings() }
}
