% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_data_dict}
\alias{get_ukb_data_dict}
\title{Get UKB data dictionary}
\usage{
get_ukb_data_dict()
}
\description{
Downloads the UK Biobank data dictionary from
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.
}
