% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_db}
\alias{get_ukb_db}
\title{Download essential UKB resources as a SQLite database file}
\usage{
get_ukb_db(directory_path)
}
\arguments{
\item{directory_path}{character. The directory where \code{ukb.db} will be
downloaded to.}
}
\value{
Returns path to \code{ukb.db} as a string invisibly
}
\description{
Downloads a file called \code{ukb.db} from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{ukbwranglr_resources}
repo to the specified directory. This is an SQLite database containing the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
data dictionary and codings files}, UKB
\href{https://biobank.ctsu.ox.ac.uk/crystal/refer.cgi?id=592}{resource 592}
(lookup tables and code mappings for various clinical coding systems) and
clinical code lists from the
\href{https://github.com/spiros/chronological-map-phenotypes}{CALIBER repo}.
}
