% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{gp_clinical_to_sqlite_db}
\alias{gp_clinical_to_sqlite_db}
\title{Pre-processing function - reformat UK Biobank primary care clinical events
data}
\usage{
gp_clinical_to_sqlite_db(df)
}
\arguments{
\item{gp_clinical_df}{A dataframe}
}
\value{

}
\description{
To be used with \code{\link{file_to_sqlite_db}}. Reformats the UK Biobank
primary care clinical events dataset to match the output format for
\code{\link{get_all_diagnostic_codes_multi}}.
}
\section{Under the hood}{


  The UK Biobank primary care data lists read codes in separate columns, one
  for Read2 and one for Read3. This function reshapes the data to long format
  so that all codes are in a single column. Note that the 3 free-text 'value'
  columns are dropped.
}

\seealso{
Other Generate a UKB database.: 
\code{\link{file_to_sqlite_db}()}
}
\concept{Generate a UKB database.}
