% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ukb.R
\name{label_ukb_main}
\alias{label_ukb_main}
\title{Label a UK Biobank main dataset}
\usage{
label_ukb_main(
  ukb_main,
  data_dict,
  ukb_codings = get_ukb_codings(),
  colnames_col = "descriptive_colnames",
  max_n_labels = 30
)
}
\arguments{
\item{ukb_main}{A UK Biobank main dataset (data frame)}

\item{data_dict}{A data dictionary specific to the UKB main dataset file,
generated by \code{\link{make_data_dict}}. This can be filtered for a
subset of columns to be labelled.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{colnames_col}{The name of the column in \code{data_dict} that contains
column names matching those in \code{ukb_main}.}

\item{max_n_labels}{Coded variables with associated value labels up to (and
equal to) this threshold will be labelled using. If \code{NULL}, then all
value labels will be applied. Default value is 30.
\code{\link[haven]{labelled}} if \code{labelled} is \code{TRUE}.}
}
\value{
A data frame with labelled
}
\description{
Adds variable and value labels a UK Biobank main dataset using
\code{\link[haven]{labelled}}. Note that labels are only supported for
numeric (integer or double) and character vectors.
}
\examples{
\dontrun{
dummy_data <- dummy_main_dataset_clinical_events()

label_ukb_main(
  ukb_main = dummy_data,
  data_dict = make_data_dict(dummy_data)
)
}
}
