% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tbls_from_dbconn}
\alias{list_tbls_from_dbconn}
\title{Make a named list of \code{tbl} objects from a \code{DBIConnection} object}
\usage{
list_tbls_from_dbconn(conn)
}
\arguments{
\item{conn}{A DBIConnection object, as returned by
\code{\link[DBI]{dbConnect}}.}
}
\value{
Named list
}
\description{
A convenience function that returns a list of \code{\link[dplyr]{tbl}}
objects from a \code{DBIConnection} object, named as per the database table names.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = ":memory:")
dplyr::copy_to(con, head(iris), "iris")
dplyr::copy_to(con, head(mtcars), "mtcars")
db_tables <- list_tbls_from_dbconn(con)
}
