% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{main_dataset_diagnoses_to_sqlite_db}
\alias{main_dataset_diagnoses_to_sqlite_db}
\title{Write diagnoses to a SQLite database}
\usage{
main_dataset_diagnoses_to_sqlite_db(
  df,
  table,
  db_path = "ukb.db",
  overwrite = FALSE,
  append = FALSE
)
}
\arguments{
\item{df}{A dataframe generated by one of the 'get all diagnostic codes'
functions}

\item{table}{Character. Name of table to write to.}

\item{db_path}{Character. Path to a SQLite database. A new database will be
created if this does not already exist. Default is "ukb.db".}

\item{overwrite}{Logical. If \code{TRUE}, will overwrite an existing table
with the same name. Default is \code{FALSE}}

\item{append}{Append to table if already exists. Default is `FALSE`.}
}
\description{
Writes a long format dataframe generated by one of the 'get all diagnostic
codes' functions (see 'See Also' section at bottom for a list of these) to a
SQLite database. If no database exists at \code{db_path}, then a new one will
be created.
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_all_diagnostic_codes_multi}()},
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()}
}
\concept{get all diagnostic codes}
