% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_empty_clinical_codes_list}
\alias{make_empty_clinical_codes_list}
\title{Create an empty list to be populated with diagnostic codes}
\usage{
make_empty_clinical_codes_list()
}
\value{
A named list
}
\description{
Returns a named list, where names are for different clinical coding systems.
This is to be populated with clinical codes and used with
\code{\link{extract_first_or_last_clinical_event}} and its related functions.
}
\examples{
make_empty_clinical_codes_list()
}
\seealso{
\code{\link{extract_first_or_last_clinical_event}}
}
