% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_mapping.R
\name{map_codes}
\alias{map_codes}
\title{Map clinical codes from one coding system to another}
\usage{
map_codes(
  codes,
  from,
  to,
  ukb_code_mappings = get_ukb_code_mappings(),
  codes_only = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{codes}{character. A character vector of codes to be mapped.}

\item{from}{character (scalar). Coding system that \code{codes} belong to.}

\item{to}{character (scalar). Coding system to map \code{codes} to.}

\item{ukb_code_mappings}{the \code{ukb_code_mappings} list returned by
\code{\link{get_ukb_code_mappings}}.}

\item{codes_only}{bool. If \code{TRUE} (default), return a character vector
of \emph{unique} codes. If \code{FALSE}, return a data frame of all results
including code descriptions (useful for manual validation).}

\item{quiet}{bool. Warning message if any of `codes` are not found for the
code type being mapped from.}
}
\description{
Uses the code mapping file provided by UK Biobank
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{resource 592}).
}
\details{
The values for arguments \code{from} and \code{to} must be one of
\code{read2}, \code{read3}, \code{icd9}, \code{icd10}, \code{bnf},
\code{dmd}, \code{read2_drugs} or \code{opcs4}.
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{get_child_codes}()},
\code{\link{lookup_codes}()}
}
\concept{Clinical code lookups and mappings}
