% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_dob}
\alias{mutate_dob}
\title{Estimate participant date of birth}
\usage{
mutate_dob(ukb_pheno, data_dict)
}
\value{
a vector of estimated dates of birth
}
\description{
Estimates dates of birth from year and month of birth (Field IDs
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=34}{34} and
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=52}{52} respectively) as
the first date of the month.
}
