% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my_skim}
\alias{my_skim}
\title{Quickly summarise a dataframe}
\usage{
my_skim(data, ...)
}
\arguments{
\item{data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\description{
A tidyverse-friendly summary function that summarises a dataframe by column type
}
\details{
Works with \code{dplyr::group_by()} and the pipe. See the
\code{\link[skimr]{skim}} documentation for more details. Adapts the
\code{skimr::skim()} function to include proportion counts for factor
variables
}
