% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenodataset.R
\name{pheno_data_dict}
\alias{pheno_data_dict}
\title{Generate a data dictionary from a raw UK Biobank phenotype file}
\usage{
pheno_data_dict(
  path,
  ukb_data_dict = "https://biobank.ctsu.ox.ac.uk/~bbdatan/Data_Dictionary_Showcase.tsv"
)
}
\arguments{
\item{path}{The path to a tab-delimited UKB phenotype file.}

\item{ukb_data_dict}{The path to a tab-delimited UKB data dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
website}.}
}
\value{
A data dictionary (data table) specific to the specified UKB
  phenotype. This includes a column of descriptive names. TODO: flesh this
  out
}
\description{
Generate a data dictionary from a raw UK Biobank phenotype file
}
