% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunked.R
\name{process_df_chunked}
\alias{process_df_chunked}
\title{Apply a function to a data frame in chunks}
\usage{
process_df_chunked(df, callback, chunk_size = 100, progress = TRUE, ...)
}
\arguments{
\item{df}{data frame}

\item{callback}{A callback function to call on each chunk}

\item{chunk_size}{The number of rows to include in each chunk}

\item{progress}{bool. Display a progress message stating time taken after
each chunk is processed.}

\item{...}{additional arguments passed on to \code{callback} function.}
}
\value{
a dataframe
}
\description{
May be helpful if applying \code{\link[tidyr]{pivot_longer}} to a large
dataframe where many of the values are \code{NA} and these are to be removed.
}
