% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRACATED_ukb_parse.R
\name{process_ukb_df_header}
\alias{process_ukb_df_header}
\title{Process UKB header}
\usage{
process_ukb_df_header(.ukb_df, return_header_only = TRUE)
}
\arguments{
\item{.ukb_df}{a raw ukb phenotypes dataframe}

\item{return_header_only}{if TRUE (default) will return the header as a vector, else returns the ukb_df renamed}
}
\value{
returns either the header as a vector or a renamed ukb df
}
\description{
Process UKB header
}
