% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenodataset.R
\name{read_pheno}
\alias{read_pheno}
\title{Read a UK Biobank phenotype file}
\usage{
read_pheno(
  path,
  delim = "\\t",
  data_dict,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  clean_dates = FALSE,
  clean_selected_continuous_and_integers = FALSE,
  na.strings = c("", "NA"),
  ...
)
}
\arguments{
\item{path}{The path to a UKB phenotype file.}

\item{delim}{Delimiter for the UKB phenotype file. Default is '\\t'}

\item{data_dict}{A data dictionary specific to the UKB phenotype file,
generated by \code{\link{make_data_dict}}. This can be a filtered version
of the full dictionary if only a subset of columns need to be loaded.}

\item{ukb_data_dict}{The UKB data dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame where all columns are of type \code{character}. By
default, this is downloaded from the \href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}} github repo.}

\item{clean_dates}{If \code{TRUE}, "nonsense" dates are converted to
\code{NA}. TODO add detail}

\item{clean_selected_continuous_and_integers}{If \code{TRUE}, certain special
continuous/integer codings are converted to \code{NA}. TODO add detail}

\item{na.strings}{ A character vector of strings which are to be interpreted as \code{NA} values. By default, \code{",,"} for columns of all types, including type \code{character} is read as \code{NA} for consistency. \code{,"",} is unambiguous and read as an empty string. To read \code{,NA,} as \code{NA}, set \code{na.strings="NA"}. To read \code{,,} as blank string \code{""}, set \code{na.strings=NULL}. When they occur in the file, the strings in \code{na.strings} should not appear quoted since that is how the string literal \code{,"NA",} is distinguished from \code{,NA,}, for example, when \code{na.strings="NA"}. }

\item{...}{Additional parameters are passed on to
\code{\link[data.table]{fread}} when reading in UKB phenotype file}
}
\value{
A UK Biobank phenotype dataset as a data table with human-readable
  variables labels and data values.
}
\description{
Read a UK Biobank phenotype file
}
