% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ukb.R
\name{read_ukb}
\alias{read_ukb}
\title{Read a UK Biobank main dataset file}
\usage{
read_ukb(
  path,
  delim = "\\t",
  data_dict = NULL,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  descriptive_colnames = TRUE,
  labelled = TRUE,
  max_n_labels = 30,
  na.strings = c("", "NA"),
  nrows = Inf,
  ...
)
}
\arguments{
\item{path}{The path to a UK Biobank main dataset file.}

\item{delim}{Delimiter for the UK Biobank main dataset file. Default is
\code{\\t}}

\item{data_dict}{A data dictionary specific to the UKB main dataset file,
generated by \code{\link{make_data_dict}}. To load only a selection of
columns, supply a filtered copy of this data dictionary containing only the
required variables. If \code{NULL} (default) then all fields will be read.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
 showcase}. This should be a data frame where all columns are of type
\code{character}. By default, this is downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{descriptive_colnames}{If \code{TRUE}, rename columns with longer
descriptive names derived from the UK Biobank's data dictionary 'Field'
column.}

\item{labelled}{If \code{TRUE}, apply variable and value labels using
\code{\link[haven]{labelled}}.}

\item{max_n_labels}{Coded variables with associated value labels up to (and
equal to) this threshold will be labelled using
\code{\link[haven]{labelled}} if \code{labelled} is \code{TRUE}.}

\item{na.strings}{ A character vector of strings which are to be interpreted as \code{NA} values. By default, \code{",,"} for columns of all types, including type \code{character} is read as \code{NA} for consistency. \code{,"",} is unambiguous and read as an empty string. To read \code{,NA,} as \code{NA}, set \code{na.strings="NA"}. To read \code{,,} as blank string \code{""}, set \code{na.strings=NULL}. When they occur in the file, the strings in \code{na.strings} should not appear quoted since that is how the string literal \code{,"NA",} is distinguished from \code{,NA,}, for example, when \code{na.strings="NA"}. }

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{...}{Additional parameters are passed on to either
\code{\link[data.table]{fread}} or \code{\link[haven]{read_dta}}}
}
\value{
A UK Biobank phenotype dataset as a data table with human-readable
  variables labels and data values.
}
\description{
Reads a UK Biobank main dataset file into R using either
\code{\link[data.table]{fread}} or \code{\link[haven]{read_dta}}. Optionally
renames variables with descriptive names and applies STATA-style variable and
value labels using the \code{\link[haven]{haven}} package.
}
\details{
Note that \code{na.strings} is not recognised by
\code{\link[haven]{read_dta}}. Reading in a STATA file may therefore require
careful checking for empty strings that ned converting to \code{NA}.
}
