% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_column}
\alias{recode_column}
\title{Recode values in a dataframe column}
\usage{
recode_column(df, col_to_recode, mapping_df)
}
\arguments{
\item{df}{a dataframe}

\item{col_to_recode}{Character - name of column in `df` to be recoded}

\item{mapping_df}{a dataframe with 2 columns named "old_vals" and "new_vals"}
}
\description{
Returns the input dataframe with recoded values in the specified
`col_to_recode`
}
\details{
Uses a `mapping_df` to recode the values for a selected column
(`col_to_recode`) in a dataframe (`df`). The `mapping_df` should only contain
2 columns named "old_vals" (containing values in `df[[col_to_recode]]` to be
recoded) and "new_vals" (replacement values).

A warning is generated if `mapping_df$old_vals` does not contain all values
in `df`, or if the return value has more rows than the original `df` (i.e. a
mutating join has been performed)
}
\section{Under the hood}{


  Uses \code{\link[dplyr]{left_join}} to merge `df` and `mapping_df`,
  retaining the "new_vals" in `mapping_df`.
}

