% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_column_coding_meaning_value_mapping_df}
\alias{recode_column_coding_meaning_value_mapping_df}
\title{Helper function for \code{\link{recode_ukbcol}}}
\usage{
recode_column_coding_meaning_value_mapping_df(
  field_id,
  ukb_data_dict,
  ukb_codings,
  mapping_direction = "meaning_code"
)
}
\arguments{
\item{field_id}{Character. A UK Biobank FieldID}

\item{ukb_data_dict}{The UKB data dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame where all columns are of type \code{character}. By
default, this is downloaded from the \href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}} github repo.}

\item{mapping_direction}{Character. Either "meaning_code" (default) or
"code_meaning"}
}
\description{
Generates a `mapping_df` for \code{\link{recode_ukbcol}}
}
\seealso{
Other recode UKB values: 
\code{\link{recode_column}()},
\code{\link{recode_ukbcol}()}
}
\concept{recode UKB values}
