% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{reformat_icd10_codes}
\alias{reformat_icd10_codes}
\title{Reformat ICD-10 codes}
\usage{
reformat_icd10_codes(
  icd10_codes,
  ukb_code_mappings,
  input_icd10_format = "ICD10_CODE",
  output_icd10_format = "ALT_CODE"
)
}
\arguments{
\item{icd10_codes}{character vector of ICD-10 codes}

\item{ukb_code_mappings}{the \code{ukb_code_mappings} list returned by
\code{\link{get_ukb_code_mappings}}.}

\item{input_icd10_format}{character. Must be either "ICD10_CODE" or
"ALT_CODE".}

\item{output_icd10_format}{character. Must be either "ICD10_CODE" or
"ALT_CODE".}
}
\value{
character vector of ICD-10 codes, reformatted as specified by
  \code{output_icd10_format}.
}
\description{
The lookup sheet in
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UKB resource 592}
for ICD-10 ("icd10_lkp") has a column called "ALT_CODE", which is an
alternative format for ICD-10 codes. This is the format used in the mapping
sheets for this resource, as well as in
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=41270}{Field ID 41270}.
This function converts from one format to the other.
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{get_child_codes}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()}
}
\concept{Clinical code lookups and mappings}
