% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{reformat_standardised_codelist}
\alias{reformat_standardised_codelist}
\title{Reformat a dataframe of clinical codes to work with
\code{\link{extract_first_or_last_clinical_event_multi}}}
\usage{
reformat_standardised_codelist(
  standardised_codelist,
  code_type,
  disease,
  disease_category,
  phenotype_source
)
}
\arguments{
\item{standardised_codelist}{a data frame with column names "code",
"description", "code_type".}

\item{code_type}{character (scalar). The clinical code type e.g. "read2"}

\item{disease}{character (scalar), e.g. "Secondary polycythaemia"}

\item{disease_category}{character (scalar). The subcategory of \code{disease}
that these codes belong to e.g. "Diagnosis of Secondary polycythaemia".}

\item{phenotype_source}{character (scalar), e.g. "caliber".}
}
\value{
A data frame with the following column names: 'disease',
  'description', 'category', 'code_type', 'code' and 'phenotype_source'.
}
\description{
A utility function that helps reformat the output from \code{\link{map_codes}}
or \code{\link{lookup_codes}} to work with
\code{\link{extract_first_or_last_clinical_event_multi}}. See also output
from \code{\link{generate_self_reported_diabetes_codes_df}} for an example of
the format that this function will output.
}
