% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_summaries.R
\name{rowise_min_max_date}
\alias{rowise_min_max_date}
\title{Get rowise min/max date from UKB dataset}
\usage{
rowise_min_max_date(
  ukb_pheno,
  selected_date_cols,
  new_colname,
  min_max = "pmin"
)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset. Must be a data table}

\item{selected_date_cols}{Character vector of column names}

\item{new_colname}{Name of new column to be created}

\item{min_max}{Character. Must be either "pmin" or "pmax"}
}
\value{
The UKB dataset supplied to argument \code{ukb_pheno} with an
  additional column containing the earliest/latest date across the
  \code{selected_date_cols}.
}
\description{
Wrapper around \code{pmin}/\code{pmax}. Removes 'nonsense' dates (see example
special coding on
\href{https://biobank.ctsu.ox.ac.uk/crystal/coding.cgi?id=819}{UK Biobank
website}).
}
\details{
The special date coding-IDs are 1313, 272, 586 and 819. Note: these are also
used in the linked primary care dataset.
}
