% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_pheno_summaries.R
\name{rowise_summary}
\alias{rowise_summary}
\title{Summarise a group of columns row-wise}
\usage{
rowise_summary(ukb_pheno, function_name, selected_cols, new_colname)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{function_name}{A function name as a character}

\item{selected_cols}{Character vector of column names}

\item{new_colname}{Name of new column to be created}
}
\description{
Summarise a group of columns row-wise
}
