% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runMakeDataset}
\alias{runMakeDataset}
\title{Launch a shiny app to interactively create a UK Biobank dataset}
\usage{
runMakeDataset()
}
\description{
Launches a shiny app
}
\examples{
\dontrun{
  runMakeDataset()
}
}
