% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{search_codes_by_description}
\alias{search_codes_by_description}
\title{Search for codes that match a description}
\usage{
search_codes_by_description(
  reg_expr,
  code_type,
  ukb_code_mappings = get_ukb_code_mappings(),
  ignore_case = TRUE,
  codes_only = FALSE,
  preferred_description_only = TRUE
)
}
\arguments{
\item{reg_expr}{a regular expression to search for}

\item{code_type}{character. The type of clinical code system to be searched.
Must be one of \code{read2}, \code{read3}, \code{icd9}, \code{icd10},
\code{bnf}, \code{dmd}, \code{read2_drugs} or \code{opcs4}.}

\item{ukb_code_mappings}{the \code{ukb_code_mappings} list returned by
\code{\link{get_ukb_code_mappings}}.}

\item{ignore_case}{Should case differences be ignored in the match?}

\item{codes_only}{bool. If \code{TRUE}, return a character vector
of \emph{unique} codes. If \code{FALSE} (default), return a data frame of all results
including code descriptions (useful for manual validation).}

\item{preferred_description_only}{bool. Return only preferred descriptions
for clinical codes with synonyms. Default value is \code{TRUE}.}
}
\value{
data frame by default, or a character vector of codes if
  \code{codes_only} is \code{TRUE}.
}
\description{
Returns a data frame with clinical codes that match the supplied regular
expression. Ignores case by default.
}
