% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_pheno_summaries.R
\name{summarise_rowise}
\alias{summarise_rowise}
\title{Summarise a UK Biobank phenotype dataset row-wise}
\usage{
summarise_rowise(
  ukb_pheno,
  functions,
  data_dict,
  grouping_col = "Field_FieldID",
  selected_col_groups,
  ...
)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{functions}{a character vector of function names (e.g. \code{c("sd")})}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{grouping_col}{the name of a column in \code{data_dict} that indicates
column groups to summarise}

\item{selected_col_groups}{a character vector of column groups to summarise
(i.e. can be a subset of the unique values in
\code{data_dict$grouping_col})}

\item{...}{arguments to passed on to summary functions listed in
\code{functions}}
}
\description{
Allows the user to summarise row-wise across a set of column "groups". e.g.
calculate standard deviation values for all systolic and diastolic blood
pressure readings per eid. Can also accept custom functions (e.g. a function
to count the number of non-NA values). Custom functions should include a
'...' parameter.
}
\examples{
\dontrun{

# make a custom function that counts the number of non-NA values.
# Note: include '...', otherwise an error is thrown by including na.rm = TRUE below

n_not_na <- function(x, ...) {
sum(!is.na(x))
}

# summarise all numerical columns in UKB dataset by calculating mean/sd/n_not_na,
# including argument na.rm = TRUE

summarise_test_read_pheno <- summarise_rowise(
ukb_pheno = ukb_pheno, # UKB phenotype dataset as a datatable
functions = c("mean", "sd", "n_not_na"),
data_dict = data_dict, # data dictionary generated by make_data_dict()
grouping_col = "Field_FieldID", # to summarise by FieldID
selected_col_groups = data_dict \%>\% # ...and only summarise numerical data
  dplyr::filter(ValueType \%in\% c("Continuous", "Integer")) \%>\%
  .$Field_FieldID \%>\%
  unique(),
na.rm = TRUE
)
}
}
