% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_summaries.R
\name{summarise_rowise_numerical_mean_min_max}
\alias{summarise_rowise_numerical_mean_min_max}
\title{Mutate mean/min/max values across instances for numerical UKB fields}
\usage{
summarise_rowise_numerical_mean_min_max(
  ukb_pheno,
  data_dict,
  mean_min_max = "rowMeans"
)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset. Must be a data table}

\item{data_dict}{A data dictionary specific to the UKB phenotype file,
generated by \code{\link{make_data_dict}}.}

\item{mean_min_max}{must be one of "rowMeans", "pmin" or "pmax". Default =
"rowMeans"}
}
\value{
the UKB df with additional columns with min/max/mean values across instances
  for numerical UKB fields
}
\description{
Much faster than \code{\link{summarise_rowise}}. However, can only summarise
mean, max or min values row-wise
}
