% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_events.R
\name{tidy_clinical_events}
\alias{tidy_clinical_events}
\title{Tidy clinical events data from a UK Biobank main dataset}
\usage{
tidy_clinical_events(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  clinical_events_sources = c("primary_death_icd10", "secondary_death_icd10",
    "self_report_medication", "self_report_non_cancer", "self_report_non_cancer_icd10",
    "self_report_cancer", "self_report_operation", "cancer_register_icd9",
    "cancer_register_icd10", "summary_hes_icd9", "summary_hes_icd10",
    "summary_hes_opcs3", "summary_hes_opcs4"),
  strict = TRUE,
  .details_only = FALSE
)
}
\arguments{
\item{ukb_main}{A UK Biobank main dataset.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
showcase}. This should be a data frame where all columns are of type
\code{character}.}

\item{ukb_codings}{The UKB codings file (available online at the UK Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
showcase}. This should be a data frame where all columns are of type
\code{character}.}

\item{clinical_events_sources}{A character vector of clinical events sources
to tidy. By default, all available options are included.}

\item{strict}{If \code{TRUE}, raise an error if required columns for any
clinical events sources listed in \code{clinical_events} are not present in
\code{ukb_main}. If \code{FALSE}, then a warning message will be displayed
instead. Default value is \code{TRUE}.}

\item{.details_only}{If \code{TRUE}, return a list detailing required Field
IDs}
}
\value{
A named list of clinical events data frames.
}
\description{
Data in a UK Biobank main dataset is stored in wide format i.e. a single row
of data per UK Biobank participant ('eid's). Clinical events may be
ascertained from numerous sources (e.g. self-reported medical conditions,
linked hospital records) with coded events and their associated dates
recorded across multiple columns. This function tidies this data into a
standardised long format table.
}
\details{
A named list of data frames is returned, with the names corresponding to the
data sources specified by \code{clinical_events}. Each data frame has the
following columns:

\itemize{ \item \code{eid} - participant identifier \item \code{source} - the
FieldID (prefixed by 'f') where clinical codes were extracted from. See
\code{\link{clinical_events_sources}} for further details. \item \code{index}
\itemize{
\item the corresponding instance and array (e.g. '0-1' means instance 0 and array
}
\enumerate{
\item \item \code{code} - clinical code. The type of clinical codings system
used depends on \code{source}. \item \code{date} - associated date. Note that
in cases where participants self-reported a medical condition but recorded
the date as either 'Date uncertain or unknown' or 'Preferred not to answer'
(see \href{https://biobank.ndph.ox.ac.uk/ukb/coding.cgi?id=13}{data coding
13}) then the date is set to \code{NA}.}
}
}
\section{Other notes}{


Results may be combined into a single data frame using
\code{\link[dplyr]{bind_rows}}.
}

\examples{
# dummy UKB main dataset and metadata
dummy_ukb_main <- get_ukb_dummy("dummy_ukb_main.tsv")
dummy_ukb_data_dict <- get_ukb_dummy("dummy_Data_Dictionary_Showcase.tsv")
dummy_ukb_codings <- get_ukb_dummy("dummy_Codings.tsv")

# tidy clinical events in a UK Biobank main dataset
clinical_events <- tidy_clinical_events(
  ukb_main = dummy_ukb_main,
  ukb_data_dict = dummy_ukb_data_dict,
  ukb_codings = dummy_ukb_codings
)

# returns a named list of data frames, one for each `clinical_events_source`
names(clinical_events)

clinical_events$summary_hes_icd10

# use .details_only = TRUE to return details of required Field IDs for
# specific clinical_events sources
tidy_clinical_events(.details_only = TRUE)
}
\seealso{
Other clinical events: 
\code{\link{clinical_events_sources}()},
\code{\link{example_clinical_codes}()},
\code{\link{extract_phenotypes}()},
\code{\link{make_clinical_events_db}()}
}
\concept{clinical events}
