% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_taken_message}
\alias{time_taken_message}
\title{Display time taken message}
\usage{
time_taken_message(start_time)
}
\arguments{
\item{start_time}{The start time.}
}
\value{
A message stating time taken since start time
}
\description{
Helper function for displaying time taken messages. Use
\code{\link[base]{proc.time}} at start of function and supply this as the
`start_time` parameter to this function.
}
