% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_parse.R
\name{ukb_mapping_generator}
\alias{ukb_mapping_generator}
\title{Title}
\usage{
ukb_mapping_generator(.ukb_df, .ukb_data_dict)
}
\arguments{
\item{.ukb_df}{}

\item{.ukb_data_dict}{}
}
\value{

}
\description{
Title
}
