% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_summary_cols.R
\name{ukb_mutate_dob}
\alias{ukb_mutate_dob}
\title{Mutate DOB column from a processed UKB df}
\usage{
ukb_mutate_dob(ukb_df, ukb_mapping_df)
}
\arguments{
\item{ukb_df}{a **cleaned** and **renamed** ukb phenotypes_file}

\item{ukb_mapping_df}{a ukb mapping file generated with ukb_mapping_generator()}
}
\value{
the ukb df with a new (DOB) column
}
\description{
Mutate DOB column from a processed UKB df
}
