% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_summary_cols.R
\name{ukb_mutate_numerical_means}
\alias{ukb_mutate_numerical_means}
\title{Mutate mean values across instances for numerical UKB fields}
\usage{
ukb_mutate_numerical_means(ukb_df, ukb_mapping_df)
}
\arguments{
\item{ukb_df}{a **cleaned** and **renamed** ukb phenotypes_file}

\item{ukb_mapping_df}{a ukb mapping file generated with ukb_mapping_generator()}
}
\value{
the UKB df with additional columns with mean values across instances for numerical UKB fields
}
\description{
Mutate mean values across instances for numerical UKB fields
}
