% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRACATED_ukb_parse.R
\name{ukb_parse}
\alias{ukb_parse}
\title{Parse a raw ukb df}
\usage{
ukb_parse(
  ukb_df,
  ukb_data_dict,
  ukb_codings,
  rename_header = TRUE,
  clean_dates = TRUE,
  clean_selected_continuous_and_integers = TRUE,
  reformat_header = FALSE
)
}
\arguments{
\item{ukb_df}{a raw ukb phenotypes file, all columns must be of type 'string'}

\item{ukb_data_dict}{the ukb data dictionary file, all columns must be of
type 'string'}

\item{ukb_codings}{the ukb codings file, all columns must be of type 'string'}

\item{rename_header}{if TRUE, rename columns with descriptive names}

\item{clean_dates}{if TRUE, convert 'nonsense' dates to 'NA'}

\item{clean_selected_continuous_and_integers}{if TRUE, convert special values
to 'NA', see 'ukb_mapping_generator'}

\item{reformat_header}{if TRUE, process header as for the .tab ukb data file
made for R, where headers are of the form e.g. 'f.5912.0.0'. This would be
converted to '5912-0.0'}
}
\value{
A parsed UKB df
}
\description{
Parse a raw ukb df
}
