% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukb_rename.R
\name{ukb_rename}
\alias{ukb_rename}
\title{Rename columns of a ukb df}
\usage{
ukb_rename(ukb_df, ukb_df_dict)
}
\arguments{
\item{ukb_df}{a UKB main dataset}

\item{ukb_df_dict}{a UKB main dataset specific to ukb_df created with `ukb_data_dictionary.R`}
}
\value{
a renamed ukb df
}
\description{
Rename columns of a ukb df
}
