% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenodataset.R
\name{ukb_rename_columns}
\alias{ukb_rename_columns}
\title{Mutates a column of descriptive colnames}
\usage{
ukb_rename_columns(pheno_data_dict)
}
\arguments{
\item{pheno_data_dict}{A data dictionary formed by joining the column names
  from a raw ukb pheno file with the ukb data dictionary
  (https://biobank.ctsu.ox.ac.uk/~bbdatan/Data_Dictionary_Showcase.tsv)

Helper function for \code{pheno_data_dict()}}
}
\description{
Mutates a column of descriptive colnames
}
