% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbwranglr.R
\docType{package}
\name{ukbwranglr}
\alias{ukbwranglr}
\title{ukbwranglr: a package for exploring UK Biobank data}
\description{
This package provides three categories of important functions:
}
\details{
\enumerate{
\item Reading UK Biobank data into R
\item Identifying participants with certain disease outcomes
\item Identifying participants taking certain medications
}
}
