% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_clinical_codes}
\alias{validate_clinical_codes}
\title{Validate a clinical codes data frame}
\usage{
validate_clinical_codes(
  clinical_codes,
  return_diseases_with_non_unique_codes = FALSE
)
}
\arguments{
\item{clinical_codes}{A data frame. See \code{\link{example_clinical_codes}}
for an example.}
}
\value{
Returns \code{TRUE} if all checks pass.
}
\description{
Checks whether a data frame of clinical codes is correctly formatted for use
with \code{\link{extract_phenotypes}}.
}
\details{
Checks that:

\itemize{ \item Expected column names are present \item All columns are of
type character \item No missing values are present in any column \item All
diseases (for each \code{author}) contain a unique set of clinical codes}

Note that currently this does not check whether the clinical codes themselves
are valid (i.e. whether a clinical code exists for a given coding system).
}
\examples{
validate_clinical_codes(example_clinical_codes())
}
