% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clinical_events_sources}
\alias{clinical_events_sources}
\title{Details for UK Biobank clinical events data}
\usage{
clinical_events_sources()
}
\value{
A data frame
}
\description{
Returns a data frame with details of clinical events data that is currently
processed by \code{\link{ukbwranglr}} functions.
}
\examples{
clinical_events_sources()
}
\seealso{
\code{\link{tidy_clinical_events}}
}
