% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_events.R
\name{clinical_events_sources}
\alias{clinical_events_sources}
\title{Details for UK Biobank clinical events data}
\usage{
clinical_events_sources()
}
\value{
A data frame
}
\description{
Returns a data frame with details of clinical events data that may be tidied
by \code{\link[=tidy_clinical_events]{tidy_clinical_events()}}.
}
\examples{
clinical_events_sources()
}
\seealso{
Other clinical events: 
\code{\link{example_clinical_codes}()},
\code{\link{extract_phenotypes}()},
\code{\link{make_clinical_events_db}()},
\code{\link{tidy_clinical_events}()}
}
\concept{clinical events}
