% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{derive_ethnic_background_simplified}
\alias{derive_ethnic_background_simplified}
\title{Derive simplified ethnic background}
\usage{
derive_ethnic_background_simplified(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  ethnicity_levels = c("White", "Mixed", "Asian or Asian British",
    "Black or Black British", "Chinese", "Other ethnic group"),
  .drop = FALSE,
  .details_only = FALSE
)
}
\arguments{
\item{ukb_main}{A UK Biobank main dataset.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ndph.ox.ac.uk/showcase/exinfo.cgi?src=AccessingData}{data
showcase}. This should be a data frame where all columns are of type
\code{character}.}

\item{ethnicity_levels}{The factor level order for the appended
\code{ethnic_background_simplified} column. By default, the baseline level
is set to "White" ethnicity.}

\item{.drop}{If \code{TRUE}, remove the required input columns from the
result}

\item{.details_only}{If \code{TRUE}, return a list containing details of
required input variables (Field IDs) and derived variables (new column
name, label and values/value labels).}
}
\value{
A data frame with a column called \code{ethnic_background_simplified}
(type factor).
}
\description{
Simplifies ethnic background in a UK Biobank main dataset to the main
categories for
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=21000}{Field ID 21000}.
}
\details{
Categories "Do not know" and "Prefer not to answer" are converted to
\code{NA}. A new column called \code{ethnic_background_simplified} of type
factor is appended to the input data frame. By default, "White" ethnicity is
set to the baseline level as this is the largest category. Levels can be
explicitly specified using the \code{ethnicity_levels} argument.
}
\examples{
library(magrittr)
# dummy UKB data and data dictionary
dummy_ukb_data_dict <- get_ukb_dummy("dummy_Data_Dictionary_Showcase.tsv")
dummy_ukb_codings <- get_ukb_dummy("dummy_Codings.tsv")

dummy_ukb_main <- read_ukb(
  path = get_ukb_dummy("dummy_ukb_main.tsv", path_only = TRUE),
  ukb_data_dict = dummy_ukb_data_dict,
  ukb_codings = dummy_ukb_codings
)

# derive ethnic background
derive_ethnic_background_simplified(
  ukb_main = dummy_ukb_main,
  ukb_data_dict = dummy_ukb_data_dict
) \%>\%
  dplyr::select(tidyselect::contains("ethnic"))
}
