% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_events.R
\name{example_clinical_codes}
\alias{example_clinical_codes}
\title{Generate an example data frame of clinical codes for diabetes}
\usage{
example_clinical_codes()
}
\value{
A data frame of selected clinical codes (non-exhaustive) for diabetes.
}
\description{
Data frames of this format may be used with \code{\link{extract_phenotypes}}.
Use \code{\link{validate_clinical_codes}} to check whether a clinical codes
data frame meets all requirements.
}
\examples{
example_clinical_codes()
}
\seealso{
\code{\link{validate_clinical_codes}}

Other clinical events: 
\code{\link{clinical_events_sources}()},
\code{\link{extract_phenotypes}()},
\code{\link{make_clinical_events_db}()},
\code{\link{tidy_clinical_events}()}
}
\concept{clinical events}
