% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{extract_first_or_last_clinical_event}
\alias{extract_first_or_last_clinical_event}
\title{Get the earliest record date for a set of diagnostic codes}
\usage{
extract_first_or_last_clinical_event(
  df,
  codes,
  min_max = "min",
  phenotype_name = "event"
)
}
\arguments{
\item{df}{Either a dataframe or a \code{\link[dplyr]{tbl}} object generated
by one of the \code{get_XXX_diagnoses} functions}

\item{codes}{Either a character vector or named list of diagnostic codes. If
a list, then each named item should be a character vector of codes (use
\code{\link{make_empty_clinical_codes_list}}).}

\item{min_max}{can be either "min" or "max". Extracts either the earliest or
latest date that on of the specified \code{codes} appears.}

\item{phenotype_name}{character. Name}
}
\value{
Data frame with columns "eid", "event_min/max_indicator" and
  "event_min/max_date". "Event" is replaced with the string supplied to the
  \code{phenotype_name} argument.
}
\description{
For a set of diagnostic codes, extract individuals with at least one of these
and retrieve, if available, the earliest or latest recorded date.
}
\section{Under the hood}{


  Filters a long format dataframe of diagnostic codes generated by one of the
  'get all diagnostic codes' functions (e.g.
  \code{\link{get_hes_icd10_diagnoses}}), filters this for the specified
  \code{codes} then groups by eid. For each eid the earliest or latest date
  is then retrieved, as specified by \code{min_max}.
}

\seealso{
Other extract specific diagnostic codes functions: 
\code{\link{extract_first_or_last_clinical_event_multi}()},
\code{\link{extract_single_diagnostic_code_record_basis}()}
}
\concept{extract specific diagnostic codes functions}
