% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{extract_first_or_last_clinical_event_multi}
\alias{extract_first_or_last_clinical_event_multi}
\title{Get the earliest record date for multiple phenotypes}
\usage{
extract_first_or_last_clinical_event_multi(
  df,
  clinical_codes_df,
  min_max = "min",
  prefix = NULL
)
}
\arguments{
\item{df}{Either a dataframe or a \code{\link[dplyr]{tbl}} object generated
by one of the \code{get_XXX_diagnoses} functions}

\item{clinical_codes_df}{data frame. Must match the format as per
\code{\link{generate_self_reported_diabetes_codes_df}}.}

\item{min_max}{can be either "min" or "max". Extracts either the earliest or
latest date that on of the specified \code{codes} appears.}

\item{prefix}{character. Optionally add a prefix to column names.}
}
\value{
Data frame with an "eid" column, and "event_min/max_indicator" and
  "event_min/max_date" columns for each phenotype in the 'category' column of
  \code{clinical_codes_df}.
}
\description{
Builds on \code{\link{extract_first_or_last_clinical_event}}, extracting
either the first or last recorded clinical event for multiple phenotypes.
}
\seealso{
Other extract specific diagnostic codes functions: 
\code{\link{extract_first_or_last_clinical_event}()},
\code{\link{extract_single_diagnostic_code_record_basis}()}
}
\concept{extract specific diagnostic codes functions}
