% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{extract_phenotypes}
\alias{extract_phenotypes}
\title{Get the earliest record date for multiple phenotype categories}
\usage{
extract_phenotypes(
  clinical_events,
  clinical_codes,
  min_max = "min",
  prefix = NULL,
  workers = NULL
)
}
\arguments{
\item{clinical_codes}{data frame. Must match the format as per
\code{\link{example_clinical_codes}}.}

\item{min_max}{can be either "min" or "max". Extracts either the earliest or
latest date that on of the specified \code{codes} appears.}

\item{prefix}{character. Optionally add a prefix to column names.}

\item{workers}{integer. The number of processes to run in parallel when
\code{clinical_codes} includes multiple diseases. This is used to set
the number of workers in \code{\link[future]{plan}} with \code{strategy =
\link[future]{multisession}}, which is passed on to
\code{\link[furrr]{future_map}}. Default value is \code{NULL}.}
}
\value{
A named list of data frames, one for each disease. Each data frame
  has an "eid" column, and "event_min/max_indicator" and "event_min/max_date"
  columns for each phenotype in the 'category' column of
  \code{clinical_codes} for that disease.
}
\description{
Builds on \code{\link{extract_first_or_last_clinical_event}}, extracting
either the first or last recorded clinical event for multiple phenotype
categories.
}
