% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{extract_single_diagnostic_code_record_basis}
\alias{extract_single_diagnostic_code_record_basis}
\title{Extract a single record from a dataframe generated by one of the
\code{get_XXX_diagnoses} functions (generic)}
\usage{
extract_single_diagnostic_code_record_basis(df, codes, mapping_function, ...)
}
\arguments{
\item{df}{Either a dataframe or a \code{\link[dplyr]{tbl}} object generated
by one of the \code{get_XXX_diagnoses} functions}

\item{codes}{Either a character vector or named list of diagnostic codes. If
a list, then each named item should be a character vector of codes (use
\code{\link{make_empty_clinical_codes_list}}).}

\item{mapping_function}{a function that will be applied to \code{df} nested
by eid, extracting a single row per eid (e.g. the earliest date that one of
the \code{codes} appears). This will be applied using
\code{\link[purrr]{map_chr}} and should return the \code{source},
\code{code} and \code{date} columns as a single character, separated by
"_SEP_".}

\item{...}{additional args are passed on to \code{mapping_function}.}
}
\value{
A dataframe nested by eid, with \code{source}, \code{code} and
  \code{date} columns for row extracted by the \code{mapping_function}.
}
\description{
Generic helper function. For a set of diagnostic codes, filter for eids with
at least one of these and extract a single row per eid (e.g. the earliest
date that one of the diagnostic codes appears).
}
\seealso{
Other extract specific diagnostic codes functions: 
\code{\link{extract_first_or_last_clinical_event_multi}()},
\code{\link{extract_first_or_last_clinical_event}()}
}
\concept{extract specific diagnostic codes functions}
