% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{field_id_pivot_longer}
\alias{field_id_pivot_longer}
\title{Convert column for a FieldID to long format}
\usage{
field_id_pivot_longer(ukb_pheno, field_id, data_dict, ukb_codings)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{field_id}{Character. A UK Biobank Field ID}

\item{data_dict}{A data dictionary specific to the UKB phenotype file,
generated by \code{\link{make_data_dict}}. This can be a filtered version
of the full dictionary if only a subset of columns need to be loaded.}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}
}
\value{
A dataframe with 3 columns: \itemize{ \item eid \item column names
  (labelled as the specified `field_id`, prefixed by 'f') \item column values
  (labelled as he specified `field_id`, prefixed by 'f' and suffixed by
  '_value') }
}
\description{
Helper function. Selects the columns relating to a specified FieldID (plus the 'eid' column)
and applies \code{\link[tidyr]{pivot_longer}}.
}
\seealso{
Other extract disease outcomes helpers: 
\code{\link{field_id_pivot_longer_multi}()}
}
\concept{extract disease outcomes helpers}
