% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{field_id_pivot_longer_multi}
\alias{field_id_pivot_longer_multi}
\title{Apply pivot_longer across multiple sets of columns}
\usage{
field_id_pivot_longer_multi(field_ids, ukb_pheno, data_dict, ukb_codings)
}
\arguments{
\item{field_ids}{Character vector of FieldIDs}

\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{A data dictionary specific to the UKB phenotype file,
generated by \code{\link{make_data_dict}}.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.}
}
\value{
Dataframe.
}
\description{
Helper function.
}
\details{
This should only be applied to related FieldIDs which are each associated
with the same number of columns. For example, FieldIDs 20002, 20008, and
20009 all pertain to self-reported non-cancer illnesses.

Raises and error if the supplied FieldIDs are associated with differing
numbers of columns, or if the result has more rows than expected.
}
\section{Under the hood}{


  Maps \code{\link{field_id_pivot_longer}} to a set of FieldIDs, then
  recombines the results into a single dataframe using
  \code{\link[dplyr]{inner_join}}.
}

\seealso{
Other extract disease outcomes helpers: 
\code{\link{field_id_pivot_longer}()}
}
\concept{extract disease outcomes helpers}
