% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_all_diagnostic_codes_multi}
\alias{get_all_diagnostic_codes_multi}
\title{Get all diagnostic codes from multiple data sources}
\usage{
get_all_diagnostic_codes_multi(
  function_list = list(get_self_report_non_cancer_diagnoses_icd10,
    get_hes_icd9_diagnoses, get_hes_icd10_diagnoses, get_cancer_register_icd9_diagnoses,
    get_cancer_register_icd10_diagnoses),
  ukb_pheno,
  data_dict,
  ukb_codings
)
}
\arguments{
\item{function_list}{A list of `get_XXXX_diagnoses` functions.}

\item{ukb_pheno}{a UK Biobank phenotype dataset}

\item{data_dict}{a data dictionary for \code{ukb_pheno}}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}
}
\value{
Dataframe
}
\description{
Extract diagnostic codes from multiple sources
}
\details{
Loops through functions from the 'get all diagnostic codes' family and
combines the results into a single dataframe.
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()}
}
\concept{get all diagnostic codes}
