% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colnames_for_fieldids}
\alias{get_colnames_for_fieldids}
\title{Get descriptive colnames associated with one or more FieldIDs}
\usage{
get_colnames_for_fieldids(field_ids, data_dict, scalar_output = FALSE)
}
\arguments{
\item{field_ids}{Character. A UK Biobank Field ID}

\item{data_dict}{a data dictionary generated by \code{\link{pheno_data_dict}}}

\item{scakar_output}{bool. If \code{TRUE}, error raised if more than one
colname is returned. Default value is \code{FALSE}.}
}
\value{
A character vector of descriptive colnames
}
\description{
Returns all descriptive column names matching one or more FieldIDs.
}
\details{
Use this in functions which manipulate a UKB phenotype dataset processed by
\code{\link{read_pheno}}. Should the convention for descriptive column names
change then so will these functions, however changes would only need to be
updated at the start of each function.
}
\seealso{
Other data dictionary helper functions: 
\code{\link{filter_data_dict}()}
}
\concept{data dictionary helper functions}
