% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_death_data_icd10_diagnoses}
\alias{get_death_data_icd10_diagnoses}
\title{Get death data ICD10 diagnoses}
\usage{
get_death_data_icd10_diagnoses(ukb_pheno, data_dict, ukb_codings)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset. Must be a data table}

\item{data_dict}{a data dictionary for \code{ukb_pheno}. Can be a "filtered"
version including only the columns to be summarised.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.}
}
\value{
Dataframe
}
\description{
Returns a long format dataframe with all death data ICD10 codes for each UK
Biobank participant.
}
\details{
Reformats the data for FieldIDs 40001 and 40002 (death register
health-related outcomes - primary and secondary causes of death; see
\href{https://biobank.ndph.ox.ac.uk/showcase/label.cgi?id=100093}{category
100093}).
}
\section{Under the hood}{


  Converts the data from FieldIDs 40001 and 40002 to long format separately
  before combining into a single dataframe.
}

\seealso{
Other get all diagnostic codes: 
\code{\link{get_all_diagnostic_codes_multi}()},
\code{\link{get_cancer_register_icd10_diagnoses}()},
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_diagnoses_set_index_code_date_cols}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_hes_opcs3_operations}()},
\code{\link{get_hes_opcs4_operations}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()},
\code{\link{main_dataset_diagnoses_to_sqlite_db}()}
}
\concept{get all diagnostic codes}
