% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{get_diagnoses_set_index_code_date_cols}
\alias{get_diagnoses_set_index_code_date_cols}
\title{Standardise a long format diagnoses dataframe}
\usage{
get_diagnoses_set_index_code_date_cols(
  get_diagnoses_df,
  index_col,
  code_col,
  date_col
)
}
\arguments{
\item{get_diagnoses_df}{a dataframe created by a 'get_X_diagnoses' function}

\item{index_col}{Column indicating diagnosis source (e.g. HES, self-report)}

\item{code_col}{Column of codes}

\item{date_col}{Column of dates when code was recorded}
}
\description{
Helper function
}
\details{
Selects essential columns: eid, field_id, code and date recorded.
}
\seealso{
Other get all diagnostic codes: 
\code{\link{get_cancer_register_icd9_diagnoses}()},
\code{\link{get_death_data_icd10_diagnoses}()},
\code{\link{get_hes_icd10_diagnoses}()},
\code{\link{get_hes_icd9_diagnoses}()},
\code{\link{get_self_report_cancer_diagnoses}()},
\code{\link{get_self_report_non_cancer_diagnoses_icd10}()},
\code{\link{get_self_report_non_cancer_diagnoses}()}
}
\concept{get all diagnostic codes}
