% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_codings}
\alias{get_ukb_codings}
\title{Download UKB codings file}
\usage{
get_ukb_codings()
}
\description{
Downloads the UK Biobank codings list from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UK
Biobank website} and reads into R with all columns as character type.
}
