% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_codings}
\alias{get_ukb_codings}
\title{Get UKB codings file}
\usage{
get_ukb_codings()
}
\description{
Downloads the UK Biobank codings dictionary from
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
github repo.
}
