% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_data_dict}
\alias{get_ukb_data_dict}
\title{Download UKB data dictionary}
\usage{
get_ukb_data_dict()
}
\description{
Downloads the UK Biobank data dictionary from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UK
Biobank website} and reads into R with all columns as character type.
}
