% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ukb_data_dict}
\alias{get_ukb_data_dict}
\title{Get the UK Biobank data dictionary}
\usage{
get_ukb_data_dict(path = NULL)
}
\arguments{
\item{path}{Either \code{NULL}, or a file path. If no file exists at this path,
then it will be downloaded directly from the UK Biobank data showcase
website to this location when the function is first called.}
}
\value{
A data frame.
}
\description{
Reads the UK Biobank data dictionary into R with all columns as type
character.
}
\details{
By default, an attempt will be made to read from a file at the path specified
by an environmental variable named \code{UKB_DATA_DICT} (see
\code{vignette('ukbwranglr')} for further details), or if this variable is not
found then the data dictionary will be downloaded directly from the \href{https://biobank.ndph.ox.ac.uk/showcase/exinfo.cgi?src=AccessingData}{UK Biobank website}
to \code{\link[=tempdir]{tempdir()}} when the function is first called.
}
\examples{
get_ukb_data_dict(
  get_ukb_dummy("dummy_Data_Dictionary_Showcase.tsv", path_only = TRUE)
)
}
