% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{get_ukb_dummy}
\alias{get_ukb_dummy}
\title{Dummy UK Biobank data}
\usage{
get_ukb_dummy(file_name, path_only = FALSE)
}
\arguments{
\item{file_name}{Name of dummy dataset file.}

\item{path_only}{If \code{TRUE}, return the file path to the dummy dataset file,
otherwise if \code{FALSE} (default), read the dummy dataset into R.}
}
\value{
A data frame if \code{path_only} is \code{FALSE} (default) or a string if
\code{path_only} is \code{TRUE}.
}
\description{
Either read a dummy UK Biobank dataset into R or return the file path only.
}
\details{
The following dummy datasets are included with this package:
\itemize{
\item \code{dummy_Data_Dictionary_Showcase.tsv}: A subset of fields from the UK
Biobank data dictionary (full version available from the UK Biobank \href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data showcase website}).
\item \code{dummy_Codings.tsv}: A subset of UK Biobank data codings (full version
available from the UK Biobank \href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data showcase website}).
\item \code{dummy_ukb_main.tsv}: A dummy main UK Biobank dataset. May be read into R
with \code{\link[=read_ukb]{read_ukb()}}. Tidy clinical events fields with \code{\link[=tidy_clinical_events]{tidy_clinical_events()}}.
\item \code{dummy_gp_clinical.txt}: A dummy UK Biobank primary care clinical event
records dataset.
\item \code{dummy_gp_scripts.txt}: A dummy UK Biobank primary care prescription
records dataset.
}
}
\examples{
library(magrittr)

# available dummy datasets
dummy_datasets <- c(
  "dummy_Data_Dictionary_Showcase.tsv",
  "dummy_Codings.tsv",
  "dummy_ukb_main.tsv",
  "dummy_gp_clinical.txt",
  "dummy_gp_scripts.txt"
)

# read dummy dataset into R
get_ukb_dummy("dummy_ukb_main.tsv")

# get file path to dummy dataset
get_ukb_dummy("dummy_ukb_main.tsv", path_only = TRUE)

# read all available dummy dataset into R
dummy_datasets \%>\%
  purrr::set_names() \%>\%
  purrr::map(get_ukb_dummy, path_only = FALSE) \%>\%
  purrr::map(tibble::as_tibble)
}
