% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_db.R
\name{make_clinical_events_db}
\alias{make_clinical_events_db}
\title{Create a SQLite database with a \code{clinical_events} table}
\usage{
make_clinical_events_db(
  ukb_main_path,
  ukb_main_delim,
  gp_clinical_path,
  ukb_db_dir,
  overwrite = FALSE,
  strict = FALSE
)
}
\arguments{
\item{ukb_main_path}{character. Path to the main UKB dataset file.}

\item{gp_clinical_path}{character. Path to the UKB primary care clinical
events file (\code{gp_clinical.txt}).}

\item{ukb_db_dir}{character. Directory where \code{ukb.db} should be written
to. An error is raised if a file called \code{ukb.db} already exists here.}

\item{strict}{logical. If TRUE, create database regardless of
whether the main UKB dataset file is missing any required clinical events
fields. Default is \code{FALSE}.}
}
\description{
Creates a SQLite database called \code{ukb.db} containing all tables from
\code{\link{get_ukb_db}}, plus an additional table called
\code{clinical_events}. This is a long format table combining all clinical
events as listed in \code{ukbwranglr:::CLINICAL_EVENTS_SOURCES}.
}
