% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ukb.R
\name{make_data_dict}
\alias{make_data_dict}
\title{Generate a UK Biobank data dictionary}
\usage{
make_data_dict(ukb_main, delim = "\\t", ukb_data_dict = get_ukb_data_dict())
}
\arguments{
\item{ukb_main}{Either the path to a UKB main dataset file (character string)
or a data frame.}

\item{delim}{Delimiter for the UKB main dataset file. Default is "\\t".
Ignored if the file name ends with \code{.dta} (i.e. is a \code{STATA}
file) or if \code{ukb_main} is a data frame.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
 showcase}. This should be a data frame where all columns are of type
\code{character}. By default, this is downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}
}
\value{
A data dictionary (data frame) specific to \code{ukb_main}. This
  includes columns with descriptive column names ("descriptive_colnames") and
  the current column names ("colheaders_raw").
}
\description{
Creates a data dictionary from either a raw UK Biobank main dataset file or a
data frame if this has already been loaded into R.
}
