% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ukb.R
\name{make_data_dict}
\alias{make_data_dict}
\title{Generate a UK Biobank data dictionary}
\usage{
make_data_dict(ukb_main, delim = "auto", ukb_data_dict = get_ukb_data_dict())
}
\arguments{
\item{ukb_main}{Either the path to a UK Biobank main dataset file (character
string) or a data frame.}

\item{delim}{Delimiter for the UKB main dataset file. Default is "auto" (see
\code{\link[data.table:fread]{data.table::fread()}}). Ignored if the file name ends with \code{.dta}
(i.e. is a STATA file) or if \code{ukb_main} is a data frame.}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
showcase}. This should be a data frame where all columns are of type
\code{character}.}
}
\value{
A data dictionary (data frame) specific to \code{ukb_main}. This
includes columns with descriptive column names ("descriptive_colnames") and
the current column names ("colheaders_raw").
}
\description{
Creates a data dictionary for a raw UK Biobank main dataset, either using
file path or from a data frame if the dataset has already been loaded into R.
}
\examples{
# dummy UKB data dictionary
dummy_ukb_data_dict <- get_ukb_dummy("dummy_Data_Dictionary_Showcase.tsv")

# create data dictionary specific to UKB main dataset, either using file path
make_data_dict(
  ukb_main = get_ukb_dummy("dummy_ukb_main.tsv", path_only = TRUE),
  ukb_data_dict = dummy_ukb_data_dict
)

# ...or from data frame
make_data_dict(
  ukb_main = get_ukb_dummy("dummy_ukb_main.tsv"),
  ukb_data_dict = dummy_ukb_data_dict
)
}
