% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenodataset.R
\name{read_pheno}
\alias{read_pheno}
\title{Read a UK Biobank phenotype file}
\usage{
read_pheno(
  path,
  delim = "\\t",
  data_dict,
  ukb_data_dict = "https://biobank.ctsu.ox.ac.uk/~bbdatan/Data_Dictionary_Showcase.tsv",
  ukb_codings = "https://biobank.ctsu.ox.ac.uk/~bbdatan/Codings.tsv",
  clean_dates = TRUE,
  clean_selected_continuous_and_integers = TRUE,
  ...
)
}
\arguments{
\item{path}{The path to a tab-delimited UKB phenotype file.}

\item{delim}{Delimiter for the UKB phenotype file. Default is '\\t'}

\item{data_dict}{A data dictionary specific to the UKB phenotype file,
generated by \code{\link{make_data_dict}}. This can be a filtered version
of the full dictionary if only a subset of columns need to be loaded.}

\item{ukb_data_dict}{The path to a tab-delimited UKB data dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{clean_dates}{If \code{TRUE}, "nonsense" dates are converted to
\code{NA}. TODO add detail}

\item{clean_selected_continuous_and_integers}{If \code{TRUE}, certain special
continuous/integer codings are converted to \code{NA}. TODO add detail}

\item{...}{Additional parameters are passed on to
\code{\link[data.table]{fread}} when reading in UKB phenotype file}
}
\value{
A UK Biobank phenotype dataset as a data table with human-readable
  variables labels and data values.
}
\description{
Read a UK Biobank phenotype file
}
