% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_column_coding_meaning_value_mapping_df}
\alias{recode_column_coding_meaning_value_mapping_df}
\title{Helper function for \code{\link{recode_ukbcol}}}
\usage{
recode_column_coding_meaning_value_mapping_df(
  field_id,
  ukb_data_dict,
  ukb_codings,
  mapping_direction = "meaning_code"
)
}
\arguments{
\item{field_id}{Character. A UK Biobank FieldID}

\item{ukb_data_dict}{The path to a tab-delimited UKB data dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{mapping_direction}{Character. Either "meaning_code" (default) or
"code_meaning"}
}
\description{
Generates a `mapping_df` for \code{\link{recode_ukbcol}}
}
\seealso{
Other recode UKB values: 
\code{\link{recode_column}()},
\code{\link{recode_ukbcol}()}
}
\concept{recode UKB values}
