% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_ukbcol}
\alias{recode_ukbcol}
\title{Recode values in a UK Biobank dataframe for a single column}
\usage{
recode_ukbcol(
  df,
  col_to_recode,
  field_id,
  ukb_data_dict,
  ukb_codings,
  mapping_direction = "meaning_code"
)
}
\arguments{
\item{df}{A dataframe}

\item{col_to_recode}{Name of column in `df` to be recoded}

\item{field_id}{Character. A UK Biobank FieldID}

\item{ukb_data_dict}{The path to a tab-delimited UKB data dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{ukb_codings}{The path to a tab-delimited UKB codings dictionary. By
default, this is downloaded directly from the
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
 website}.}

\item{mapping_direction}{Character. Either "meaning_code" (default) or
"code_meaning"}
}
\description{
Recodes values in a specified column from descriptive label to UK Biobank
coding (default) or vice versa
}
\section{Under the hood}{


The UK Biobank codings for a given FieldID (`field_id`) are extracted into
  a `mapping_df`, which is formatted for use with
  \code{\link{recode_column}}. This then recodes the specified column
  (`col_to_recode`) in `df` from either meaning to raw UK Biobank codings
  (default) or vice versa.
}

\seealso{
columns "Value" and "Meaning" in the
  \href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UKB
   codings dictionary}

Other recode UKB values: 
\code{\link{recode_column_coding_meaning_value_mapping_df}()},
\code{\link{recode_column}()}
}
\concept{recode UKB values}
