% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_summaries.R
\name{summarise_numerical_variables}
\alias{summarise_numerical_variables}
\title{Summarise numerical variables}
\usage{
summarise_numerical_variables(
  ukb_main,
  data_dict = NULL,
  summary_function = "mean",
  summarise_by = "Field",
  .drop = FALSE
)
}
\arguments{
\item{ukb_main}{A UK Biobank main dataset data frame. Column names must match
those under the \code{descriptive_colnames} column in \code{data_dict}.}

\item{data_dict}{a data dictionary specific to the UKB main dataset file,
created by \code{\link{make_data_dict}}.}

\item{summary_function}{The summary function to be applied. Options: "mean",
"min", "max", "sum" or "n_values"}

\item{summarise_by}{Whether to summarise by "Field" or by "Instance".}

\item{.drop}{If \code{TRUE}, removes the original numerical variables from
the result. Default value is \code{FALSE}.}
}
\value{
A data frame
}
\description{
Summarises numerical variables with repeated measurements either by field
(i.e. all available measurements) or by instance (i.e. for all measurements
at each assessment visit). Currently available summary options are mean,
minimum, maximum, sum and number of non-missing values.
}
\details{
Note that when\code{summary_function = "sum"}, missing values are converted
to zero. Therefore if a set of values are \emph{all} missing then the sum
will summarised as \code{0}. See the documentation for
\code{\link[base]{rowSums}} for further details.
}
\examples{
# get dummy data
dummy_data_path <- system.file("extdata", "dummy_ukb_data.csv", package = "ukbwranglr")
read_ukb(dummy_data_path, delim = ",") \%>\%
  summarise_numerical_variables()
}
