% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_summaries.R
\name{summarise_numerical_variables}
\alias{summarise_numerical_variables}
\title{Summarise numerical variables}
\usage{
summarise_numerical_variables(
  ukb_main,
  data_dict,
  summary_function = "Mean",
  summarise_by = "Field",
  .drop = FALSE
)
}
\arguments{
\item{ukb_main}{Either the path to a UKB main dataset file (character string)
or a data frame.}

\item{summary_function}{The summary function to be applied. Options are
"Mean", "Min", "Max" or "Sum"}

\item{summarise_by}{Whether to summarise by "Field" or by "Instance".}

\item{.drop}{If \code{TRUE}, removes the numerical variables being summarised
from the result. Default value is \code{FALSE}}
}
\value{
A data frame
}
\description{
Summarises numerical variables with repeated measurements either by field
(i.e. all available measurements) or by instance (i.e. for all measurements
at each assessment visit). Currently available summary options are mean,
minimum, maximum and sum.
}
