% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise_summaries.R
\name{summarise_rowise}
\alias{summarise_rowise}
\title{Summarise a UK Biobank phenotype dataset row-wise}
\usage{
summarise_rowise(
  ukb_pheno,
  functions,
  data_dict,
  grouping_col = "Field_FieldID",
  prefix = NULL,
  ...
)
}
\arguments{
\item{ukb_pheno}{a UK Biobank phenotype dataset. Must be a data table}

\item{functions}{a character vector of function names (e.g. \code{c("sd")})}

\item{data_dict}{a data dictionary for \code{ukb_pheno}. Can be a "filtered"
version including only the columns to be summarised.}

\item{grouping_col}{the name of a column in \code{data_dict} that indicates
column groups to summarise}

\item{prefix}{character. An optional prefix to add to the names of all newly
created columns. Default is \code{NULL}.}

\item{...}{arguments to passed on to summary functions listed in
\code{functions}}
}
\description{
Allows the user to summarise row-wise across a set of column "groups". e.g.
calculate standard deviation values for all systolic and diastolic blood
pressure readings per eid. Can also accept custom functions (e.g. a function
to count the number of non-NA values). Custom functions should include a
'...' parameter. \strong{Note: \code{ukb_pheno} must be a data table.}
}
\examples{
\dontrun{
# make a custom function that counts the number of non-NA values.
# Note: include '...', otherwise an error is thrown by including na.rm = TRUE below

n_not_na <- function(x, ...) {
sum(!is.na(x))
}

# summarise all numerical columns in UKB dataset by calculating mean/sd/n_not_na,
# including argument na.rm = TRUE

ukb_pheno_summarised <- summarise_rowise(
  # UKB phenotype dataset as a datatable
  ukb_pheno = ukb_pheno,

  functions = c("mean", "sd", "n_not_na"),

  # data dictionary generated by make_data_dict(), filtered
  # for only numerical data
  data_dict = data_dict \%>\%
    dplyr::filter(ValueType \%in\% c("Continuous", "Integer")),

  # to summarise by FieldID
  grouping_col = "Field_FieldID"

  # additional args passed on to `functions`
  na.rm = TRUE
)
}
}
