% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_phenotypes.R
\name{tidy_clinical_events}
\alias{tidy_clinical_events}
\title{Tidy clinical events data from a UK Biobank main dataset}
\usage{
tidy_clinical_events(
  ukb_main,
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings(),
  clinical_events_sources = c("primary_death_icd10", "secondary_death_icd10",
    "self_report_non_cancer", "self_report_non_cancer_icd10", "self_report_cancer",
    "self_report_operation", "cancer_register_icd9", "cancer_register_icd10",
    "summary_hes_icd9", "summary_hes_icd10", "summary_hes_opcs3", "summary_hes_opcs4"),
  strict = TRUE,
  .details_only = FALSE
)
}
\arguments{
\item{ukb_main}{A UK Biobank main dataset. Must be a \code{\link[data.table]}}

\item{ukb_data_dict}{The UKB data dictionary (available online at the UK
Biobank
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{data
 showcase}. This should be a data frame where all columns are of type
\code{character}. By default, this is downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{ukb_codings}{The UKB codings dictionary. This should be a data frame
where all columns are of type \code{character}. By default, this is
downloaded from the
\href{https://github.com/rmgpanw/ukbwranglr_resources}{\code{ukbwranglr_resources}}
 github repo.}

\item{clinical_events_sources}{A character vector of clinical events sources to tidy.
By default, all available options are included.}

\item{strict}{If \code{TRUE}, raise an error if required columns for any
clinical events sources listed in \code{clinical_events} are not present in
\code{ukb_main}. If \code{FALSE}, then a warning message will be displayed
instead. Default value is \code{TRUE}.}

\item{.details_only}{If \code{TRUE}, return a list detailing required Field
IDs}
}
\value{
A named list of clinical events data frames.
}
\description{
Data in a UK Biobank main dataset is stored in wide format i.e. a single row
of data per UK Biobank participant ('eid's). Clinical events may be
ascertained from numerous sources (e.g. self-reported medical conditions,
linked hospital records) with coded events and their associated dates
recorded across multiple columns. This function tidies this data into a
standardised long format table.
}
\details{
A named list of data frames is returned, with the names corresponding to the
data sources specified by \code{clinical_events}. Each data frame has the
following columns:

\itemize{ \item \code{eid} - participant identifier \item \code{source} - the
FieldID (prefixed by 'f') where clinical codes were extracted from \item
\code{index} - the corresponding instance and array (e.g. '0-1' means
instance 0 and array 1) \item \code{code} - clinical code. See
\code{\link{clinical_events_sources}} for further details \item \code{date} -
associated date }

Results may be combined into a single data frame using
\code{\link[dplyr]{bind_rows}}.
}
\examples{
# tidy clinical events in a UK Biobank main dataset
tidy_clinical_events(
  ukb_main = dummy_main_dataset_clinical_events(),
  ukb_data_dict = get_ukb_data_dict(),
  ukb_codings = get_ukb_codings()
)

# use .details_only = TRUE to return details of required Field IDs for specific clinical_events sources
tidy_clinical_events(.details_only = TRUE)
}
