% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRACATED_ukb_parse.R
\name{ukb_mapping_generator}
\alias{ukb_mapping_generator}
\title{Title}
\usage{
ukb_mapping_generator(.ukb_df, .ukb_data_dict, reformat_header = TRUE)
}
\arguments{
\item{.ukb_df}{A raw ukb phenotypes file}

\item{.ukb_data_dict}{The ukb data dictionary file}

\item{reformat_header}{if TRUE, process header as for the .tab ukb data file
made for R, where headers are of the form e.g. 'f.5912.0.0'. This would be
converted to '5912-0.0'}
}
\value{
Mapping df
}
\description{
Title
}
