% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbwranglr-package.R
\docType{package}
\name{ukbwranglr-package}
\alias{ukbwranglr}
\alias{ukbwranglr-package}
\title{ukbwranglr: Exploring UKB Data}
\description{
Tools to load, pre-process and summarise UK Biobank data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rmgpanw/ukbwranglr}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com}

}
\keyword{internal}
