% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbwranglr.R
\docType{package}
\name{ukbwranglr}
\alias{ukbwranglr}
\title{ukbwranglr: a package for UK Biobank data wrangling}
\description{
This package provides three categories of helpful functions:
}
\details{
\enumerate{

\item Reading UK Biobank data into R - \code{\link{make_data_dict}} and
\code{\link{read_ukb}}

\item Summarising numerical variables -
\code{\link{summarise_numerical_variables}}

\item Identifying phenotypic outcomes - \code{\link{tidy_clinical_events}},
\code{\link{make_clinical_events_db}} and \code{\link{extract_phenotypes}}

}
}
